requirements.txt
python-telegram-bot==20.7
firebase-admin==6.4.0
python-dotenv==1.0.1